+++ lib/custom_refresh_indicator.dart
@@ -1,6 +1,5 @@
 library custom_refresh_indicator;
 
-import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';